Google Classroom Exporter: final refinements
