--- src/lib/exporters/googleClassroom/stripHtml.ts
@@ -1,14 +1,0 @@
-export function stripHtml(html: string | undefined | null): string {
-  if (!html) {
-    return ""
-  }
-
-  return html
-    .replace(/<[^>]*>/g, " ")
-    .replace(/&nbsp;/gi, " ")
-    .replace(/&amp;/gi, "&")
-    .replace(/&lt;/gi, "<")
-    .replace(/&gt;/gi, ">")
-    .replace(/\s+/g, " ")
-    .trim()
-}
